API: move class LinkageTree to module pytools.data
