+++ src/pytools/data/__init__.py
@@ -2,5 +2,6 @@
 Support for simulating classification and regression data.
 """
 
+from ._linkage import *
 from ._matrix import *
 from ._simulation import *
