+++ src/pytools/data/base/_linkage.py
