+++ src/pytools/viz/dendrogram/__init__.py
@@ -3,5 +3,4 @@ Drawer and styles for dendrogram representations of linkage trees.
 """
 
 from ._draw import *
-from ._linkage import *
 from ._style import *
